link from history row to history table

+++ cephia/cephia/file_handlers/test_history_file_handler.py
@@ -50,7 +50,8 @@ class TestHistoryFileHandler(FileHandler):
     
     def validate(self):
         from cephia.models import Subject
-        from diagnostics.models import DiagnosticTestHistoryRow, ProtocolLookup, TestPropertyEstimate
+        from diagnostics.models import (DiagnosticTestHistoryRow, ProtocolLookup,
+                                        TestPropertyEstimate, DiagnosticTestHistory)
 
         rows_validated = 0
         rows_failed = 0
@@ -82,6 +83,10 @@ class TestHistoryFileHandler(FileHandler):
                 except TestPropertyEstimate.DoesNotExist:
                     error_msg += "Property Estimate not recognised.\n"
 
+
+                if DiagnosticTestHistory.objects.filter(subject__subject_label=test_history_row.subject, ignore=True).exists():
+                    error_msg += "Cannot overwrite test history data that has already been edited.\n"
+                    
                 if error_msg:
                     raise Exception(error_msg)
 
@@ -111,7 +116,10 @@ class TestHistoryFileHandler(FileHandler):
                 excluded_subjects = DiagnosticTestHistoryRow.objects.values_list('subject', flat=True).filter(fileinfo=self.upload_file, state='error').distinct()
                 
                 for subject_label in excluded_subjects:
-                    DiagnosticTestHistory.objects.filter(subject__subject_label=subject_label).delete()
+                    if not DiagnosticTestHistory.objects.filter(subject__subject_label=subject_label, ignore=True).exists():
+                        DiagnosticTestHistory.objects.filter(subject__subject_label=subject_label).delete()
+                    else:
+                        continue
                     
                 for subject_row_error in DiagnosticTestHistoryRow.objects.filter(subject__in=excluded_subjects, state='validated', fileinfo=self.upload_file):
                     subject_row_error.state = 'error'
@@ -134,6 +142,11 @@ class TestHistoryFileHandler(FileHandler):
                         if not test_history.subject.subject_eddi:
                             test_history.subject.subject_eddi = SubjectEDDI.objects.create()
                             test_history.subject.save()
+                        else:
+                            test_history.subject.subject_eddi.tci_begin = None
+                            test_history.subject.subject_eddi.tci_end = None
+                            test_history.subject.subject_eddi.tci_size = None
+                            test_history.subject.subject_eddi.eddi = None
 
                         test_history.subject.subject_eddi.recalculate = True
                         test_history.subject.subject_eddi.save()
@@ -144,6 +157,7 @@ class TestHistoryFileHandler(FileHandler):
                         test_history_row.state = 'processed'
                         test_history_row.error_message = ''
                         test_history_row.date_processed = timezone.now()
+                        test_history_row.test_history = test_history
                         test_history_row.save()
 
                         rows_inserted += 1
